support both http and ws providers
